+++ examples/echo_bot.py
@@ -1,13 +1,17 @@
-"""
-It answers "pong" if receives "ping"
-"""
-
 import logging
 import argparse
 
 from telegram.client import Telegram
 
 from utils import setup_logging
+
+
+"""
+It answers "pong" if receives "ping"
+
+Usage:
+    python examples/send_message.py api_id api_hash phone
+"""
 
 
 if __name__ == '__main__':
@@ -23,6 +27,7 @@ if __name__ == '__main__':
         api_id=args.api_id,
         api_hash=args.api_hash,
         phone=args.phone,
+        database_encryption_key='changeme1234',
     )
     # you must call login method before others
     tg.login()